feat(async): decorate _update_attrs with awaitable_postprocess
Since _update_attrs is some sort of postprocess then we decorated it so
Hence changed interface we now can explicitly user self._update_attrs
but be sure that we return it, since it's awaitable
